# docs_caich/annotation_openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Construction Safety Annotation API",
    "description": "Free-form GroundingDINO detection, SAM box-prompt segmentation, Qwen prompt enrichment, manual annotation review, and dataset release API. The API process does not load model weights.",
    "version": "1.2.0"
  },
  "paths": {
...
    "/v1/annotation/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create Annotation Job",
        "operationId": "createAnnotationJob",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 8,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Asset was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Job request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "429": {
            "description": "Job queue is full.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Annotation Job",
        "operationId": "getAnnotationJob",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel Annotation Job",
        "operationId": "cancelAnnotationJob",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "409": {
            "description": "Job is already in a terminal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/jobs/{job_id}/detections": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List Annotation Job Detections",
        "operationId": "listAnnotationJobDetections",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetectionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/jobs/{job_id}/review-tasks": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Build Detection Tasks",
        "operationId": "buildDetectionReviewTasks",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildDetectionTasksRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildReviewTasksResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "409": {
            "description": "Detection job has not completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Detection selection is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/jobs/{job_id}/assets/{asset_id}/bbox-image": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Annotation Job Bbox Image",
        "operationId": "getAnnotationJobBoundingBoxImage",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image with GroundingDINO bounding boxes.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildDetectionTasksRequest": {
        "properties": {
          "detection_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 500,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Ids"
          },
          "category": {
            "$ref": "#/components/schemas/AnnotationCategory",
            "default": "unsafe"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BuildDetectionTasksRequest"
      },
      "BuildReviewTasksResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Task Ids"
          },
          "created_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Created Count"
          },
          "existing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Existing Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "created_count",
          "existing_count"
        ],
        "title": "BuildReviewTasksResponse"
      },
      "CancelJobRequest": {
        "properties": {
          "actor_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actor_id",
          "reason"
        ],
        "title": "CancelJobRequest"
      },
...
      "CreateJobRequest": {
        "properties": {
          "asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Asset Ids"
          },
          "grounding_prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Grounding Prompt"
          },
          "grounding_prompt_normalization_mode": {
            "type": "string",
            "enum": [
              "off",
              "terminal_period",
              "canonical_terms",
              "llm_grounding_caption"
            ],
            "title": "Grounding Prompt Normalization Mode",
            "default": "terminal_period"
          },
          "grounding_prompt_normalization_profile": {
            "type": "string",
            "enum": [
              "construction_safety_v1",
              "open_semantic_zh_en_v1"
            ],
            "title": "Grounding Prompt Normalization Profile",
            "default": "construction_safety_v1"
          },
          "grounding_prompt_translation_failure_policy": {
            "type": "string",
            "enum": [
              "fail_job",
              "fallback_canonical_terms",
              "fallback_terminal_period"
            ],
            "title": "Grounding Prompt Translation Failure Policy",
            "default": "fallback_canonical_terms"
          },
          "pipeline_version": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Pipeline Version",
            "default": "groundingdino-free-form-v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "asset_ids",
          "grounding_prompt"
        ],
        "title": "CreateJobRequest"
      },
...
      "DetectionJobProgress": {
        "properties": {
          "total_assets": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Assets"
          },
          "completed_assets": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completed Assets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_assets",
          "completed_assets"
        ],
        "title": "DetectionJobProgress"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_error",
          "annotation_validation_failed",
          "unauthorized",
          "forbidden",
          "not_found",
          "version_conflict",
          "idempotency_conflict",
          "invalid_state_transition",
          "unsupported_media_type",
          "request_too_large",
          "queue_full",
          "model_unavailable",
          "downstream_unavailable",
          "internal_error"
        ],
        "title": "ErrorCode"
      },
...
      "ErrorPayload": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "const": "grounding_dino"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "grounding_prompt": {
            "type": "string",
            "title": "Grounding Prompt"
          },
          "grounding_prompt_normalization_mode": {
            "type": "string",
            "enum": [
              "off",
              "terminal_period",
              "canonical_terms",
              "llm_grounding_caption"
            ],
            "title": "Grounding Prompt Normalization Mode",
            "default": "terminal_period"
          },
          "grounding_prompt_normalization_profile": {
            "type": "string",
            "enum": [
              "construction_safety_v1",
              "open_semantic_zh_en_v1"
            ],
            "title": "Grounding Prompt Normalization Profile",
            "default": "construction_safety_v1"
          },
          "grounding_prompt_translation_failure_policy": {
            "type": "string",
            "enum": [
              "fail_job",
              "fallback_canonical_terms",
              "fallback_terminal_period"
            ],
            "title": "Grounding Prompt Translation Failure Policy",
            "default": "fallback_canonical_terms"
          },
          "progress": {
            "$ref": "#/components/schemas/DetectionJobProgress"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StageResult"
            },
            "propertyNames": {
              "const": "grounding_dino"
            },
            "type": "object",
            "title": "Stages"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/JobError"
            },
            "type": "array",
            "title": "Errors"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "status",
          "pipeline_version",
          "grounding_prompt",
          "progress",
          "stages",
          "created_at"
        ],
        "title": "Job"
      },
...
      "JobDetectionsResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobDetection"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "total"
        ],
        "title": "JobDetectionsResponse"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial_failed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}